Moved predicates from utils/sync.py and utils/dataclass.py to utils.predicates.py for consistency
