+++ litestar/utils/__init__.py
@@ -1,13 +1,15 @@
 from litestar.utils.deprecation import deprecated, warn_deprecation
 
-from .dataclass import is_dataclass_class, is_dataclass_instance
 from .helpers import Ref, get_enum_string_value, get_name
 from .path import join_paths, normalize_path
 from .predicates import (
     is_any,
+    is_async_callable,
     is_attrs_class,
     is_class_and_subclass,
     is_class_var,
+    is_dataclass_class,
+    is_dataclass_instance,
     is_generic,
     is_mapping,
     is_non_string_iterable,
@@ -26,7 +28,7 @@ from .scope import (
     set_litestar_scope_state,
 )
 from .sequence import compact, find_index, unique
-from .sync import AsyncCallable, AsyncIteratorWrapper, async_partial, is_async_callable
+from .sync import AsyncCallable, AsyncIteratorWrapper, async_partial
 from .typing import annotation_is_iterable_of_type, get_origin_or_inner_type, make_non_optional_union
 
 __all__ = (
