+++ litestar/utils/predicates.py
@@ -1,11 +1,16 @@
 from __future__ import annotations
 
+from asyncio import iscoroutinefunction
 from collections import defaultdict, deque
 from collections.abc import Iterable as CollectionsIterable
+from dataclasses import is_dataclass
+from functools import partial
 from inspect import isasyncgenfunction, isclass, isgeneratorfunction
 from typing import (
     TYPE_CHECKING,
     Any,
+    Awaitable,
+    Callable,
     ClassVar,
     DefaultDict,
     Deque,
@@ -37,6 +42,8 @@ from litestar.utils.typing import get_origin_or_inner_type
 if TYPE_CHECKING:
     from litestar.types.builtin_types import TypedDictClass
     from litestar.types.callable_types import AnyGenerator
+    from litestar.types.protocols import DataclassProtocol
+
 
 try:
     import pydantic
@@ -49,10 +56,13 @@ except ImportError:  # pragma: no cover
     attrs = Empty  # type: ignore
 
 __all__ = (
+    "is_async_callable",
     "is_any",
     "is_attrs_class",
     "is_class_and_subclass",
     "is_class_var",
+    "is_dataclass_class",
+    "is_dataclass_instance",
     "is_generic",
     "is_mapping",
     "is_non_string_iterable",
@@ -68,6 +78,51 @@ __all__ = (
 
 P = ParamSpec("P")
 T = TypeVar("T")
+
+
+def is_async_callable(value: Callable[P, T]) -> TypeGuard[Callable[P, Awaitable[T]]]:
+    """Extend :func:`asyncio.iscoroutinefunction` to additionally detect async :func:`functools.partial` objects and
+    class instances with ``async def __call__()`` defined.
+
+    Args:
+        value: Any
+
+    Returns:
+        Bool determining if type of ``value`` is an awaitable.
+    """
+    while isinstance(value, partial):
+        value = value.func  # type: ignore[unreachable]
+
+    return iscoroutinefunction(value) or (
+        callable(value) and iscoroutinefunction(value.__call__)  #  type: ignore[operator]
+    )
+
+
+def is_dataclass_instance(obj: Any) -> TypeGuard[DataclassProtocol]:
+    """Check if an object is a dataclass instance.
+
+    Args:
+        obj: An object to check.
+
+    Returns:
+        True if the object is a dataclass instance.
+    """
+    return hasattr(type(obj), "__dataclass_fields__")
+
+
+def is_dataclass_class(annotation: Any) -> TypeGuard[type[DataclassProtocol]]:
+    """Wrap :func:`is_dataclass <dataclasses.is_dataclass>` in a :data:`typing.TypeGuard`.
+
+    Args:
+        annotation: tested to determine if instance or type of :class:`dataclasses.dataclass`.
+
+    Returns:
+        ``True`` if instance or type of ``dataclass``.
+    """
+    try:
+        return isclass(annotation) and is_dataclass(annotation)
+    except TypeError:  # pragma: no cover
+        return False
 
 
 def is_class_and_subclass(annotation: Any, t_type: type[T]) -> TypeGuard[type[T]]:
